# Rework contribution score function backend