## Background
When trying to explain even one sequence from Borzoi, the explaining functions would always OOM. Digging a bit revealed `integrated_grad` makes a batch of 26 sequences to predict it and the background, which is never gonna work for a massive model like this. There was a preliminary batching function in the explainer code, but it wasn't accessible from the CREsted wrapper, and (at least on TensorFlow) its implementation gave me a lot of trouble. While fixing this, a combination of chasing speedups and frustration with code duplication led to a big rework of the explainer functions. 

## Main changes
### Structure
I cleaned up the function structure, separating only a few 'primitive' functions that operate on one sequence/batch to directly calculate gradients (and the batching function `function_batch()`) as backend-specific functions. The rest of the functions, which do more high-level creating backgrounds/averaging over gradients/etc are shared between tensorflow and pytorch.
To do this batching, I now reworked the batching function, added it to all functions, and exposed `batch_size` from CREsted's `tl.contribution_scores()`. 

### Integrated gradients
Default code should work exactly the same for 'integrated_grad': there, it's always processed every sequence individually and turned it into size 26 batch inside every `integrated_grad()` call, which is far under the default 128 batch size. If it's under the batch size, the function is called without any batching. I considered integrating this over sequences so that we can batch multiple sequences in one run, but that led to some bugs, and would possibly lead to problems on the expected integrated gradient side (saving 26*500 gradients might still work, but saving 26*25*500=325000 could very well cause memory issues, and then we'd need CPU-level batching as well, and blegh. Integrated gradients are plenty fast already.)

### Expected integrated gradients
For 'expected_integrated_grad', it previously called `integrated_grad()` 25 times (once for each baseline). I've simplified this and integrated it in the default `integrated_grad()` function. Therefore, instead of 25 calls of 26 sequences ( `n_steps+1` sequences for `n_baselines` baselines) it now creates 650 sequences (`n_baselines`*(`n_steps+1`)) in one go and only needs 6 batches of 128 seqs.
Also, previously, the shuffling didn't accept a seed value. I've changed that so that expected integrated gradients should be more consistent now.

## Other mini changes
- I exposed saliency maps (aka raw gradients) from `tl.contribution_scores()` - the recommended technique should still be `expected_integrated_grad` but I can see people need this since it's way way faster especially when you're working with low batch sizes (since expected_integrated_grad takes 25*26=650 sequence explanations to explain one sequence, which hurts when that's 650 gradient tracking predictions instead of 25 batches of 26). More selfishly, I needed it for the Borzoi comparisons, and it's implemented anyway so might as well expose it. 
- Also, I added mutagenesis to the test suite - if it's too slow we can remove it again, but it is an important function that wasn't being tested.

## Review
- I haven't tested this on PyTorch - I trust the CI will catch anything wrong, but an extra look over the code on that side is especially welcome

## Further changes
- A massive speedup would be to implement Seppe's idea of getting gradients for multiple classes in the `saliency_map` function, rather than running everything again per class - I don't have time to implement that atm. 
- Mutagenesis code probably deserves a look, I didn't touch that beyond making it accept batch_size atm.