@casblaauw 2025-03-09
Note: this is related to #104 , just tagging it here. This could be a way to fix the OOM issues there, or get both speedups and keep memory load managable?

Also, I have zero clue why it's failing on a loss function not being serialised, I didn't touch any of that. Problem for tomorrow.