@casblaauw 2025-03-13
Some speed and equivalence tests:

Raw saliency maps:
(Compares with prev also-batched implementation, which gave a warning every time it's used)
```
DilatedCNN model (500 seqs): old 0.69s vs new 0.69s
DilatedCNN model (500 seqs) old vs new saliencymap equal: True, allclose True
```

Integrated gradients:
```
DilatedCNN model (3 seqs): old 0.05s vs new 0.05s
DilatedCNN model (3 seqs) old new integrated grads: equal: False, allclose atol 1e-6 True
DilatedCNN model (500 seqs): old 26.22s vs new 25.31s
Borzoi model (3 seqs): old integrated gradients impossible vs new integrated grads 29.04s
```

Expected integrated gradients
```
DilatedCNN model (3 seqs): old exp int grads 3.91 vs new exp int grads 2.82
DilatedCNN model (100 seqs): old exp int grads 130.75 vs new exp int grads 94.12
DilatedCNN model (1 seq) old exp int grads vs new exp int grads (batch size 128) equal: False, allclose atol 1e-05 True
```

Some of the sources of variance:
- In expected integrate gradients, sequence inputs were exactly the same, but running them in 2 batches of 26 vs 1 batch of 52 gave ever so slightly different gradients. 1e-7 level stuff, so not detectable in practice.
- Similarly (but even more minor), saving arrays to a list and concatenating vs writing them to a slice of a pre-made array led to miniscule differences.
For each of these 3, I checked the results from explaining one sequence, and they look exactly the same when plotted. The previous variance of setting a seed for expected integrated gradients will have been much higher.