@casblaauw 2025-03-13
My tests are failing on an unrelated issue to the PR: `CosineMSELogLoss` not being decorated with `@keras.saving.register_keras_serializable(package="Losses")`, which means a model can't be loaded in with compiled=True if it used that loss. While it's true that that function doesn't have that decorator, unlike `CosineMSELoss`, this has always been the case... The test that's failing is also at least 3 months old.

I've now added it as a commit, no clue why it decided to complain now and not in the past few months but okay.