@casblaauw 2025-03-13
Tagging @LukasMahieu and @SeppeDeWinter for a look over this, if at all possible also testing the function on pytorch if you have an environment ready.