@casblaauw 2025-03-14
I've added type annotation and more docs, you're right - it's a small hassle, and at least adding them every time we go over a function will increase coverage over time.

For the transfers to/from the GPU, I did originally write this to minimise GPU memory, but handling the entire matrix in GPU (i.e. `(pred(batch) for batch in batches).numpy()` vs `(pred(batches).numpy() for batch in batches)`) would probably be a bit of a speedup. I do still like transferring to CPU/numpy somewhat quickly, simply because almost any tensor-handling code will need to be duplicated for tensorflow and pytorch, which is horrible to maintain.

Okay, I tried a quick implementation, but a few tests show that it doesn't really change anything (I think):
```
# Original CREsted implementation 
Integrated grads, 500 seqs: old old 26.22s
Expected integrated grads, 100 seqs: old old 131.55s

# Current branch: move each batch separately and insert in array
Integrated grads, 500 seqs: move each batch, write to numpy array: 25.30s
Expected integrated grads, 100 seqs: move each batch, write to numpy array: 91.22

# New: save to list, concat with tf.concat, then move to numpy:
Integrated grads, 500 seqs: concatenating list before transferring: 25.26s
Expected integrated grads, 100 seqs: concatenating list before transferring: 91.39

# New: save to tf.Variable array, then move to numpy:
Integrated grads, 500 seqs: writing to tf.variable array: 25.21s
Expected integrated grads, 100 seqs:  writing to tf.variable array: 90.97
```

So I think I'll keep it as is, since that 1. reduces GPU memory load and 2. provides a clear 'separation' between shared numpy code and the backend-specific functions. It might lead to torch speedups though, and maybe doing the entire expected integrated grad integration/averaging on GPU before transferring would speed up more (transferring `(1, seq_len, 4)` instead of `(num_baselines, num_steps, seq_len, 4)`). Question for a later PR.