@casblaauw 2025-03-16
Tested both `crested.tl.contribution_scores` and underlying functions in PyTorch with the new code, works and gives equivalent results to with TensorFlow (when judged by plotting scatterplots of scores on top of eachother). Merging this.